+++ CHANGELOG.md
@@ -26,6 +26,11 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/jira-software-actions/compare/release-1.3.6...master
 
+### Added
+- New action `WorkOnBacklog` that visits the backlog and has a chance to edit an existing sprint by changing its name. Resolve [JPERF-873].
+
+### Deprecated
+- Deprecate `ViewBacklogAction` in favor of `WorkOnBacklog`
 ## [1.3.6] - 2022-10-19
 [1.3.6]: https://github.com/atlassian/jira-software-actions/compare/release-1.3.5...release-1.3.6